+++ src/View/Helper/LibraryHelper.php
@@ -13,7 +13,6 @@
 namespace MeTools\View\Helper;
 
 use Cake\Core\Plugin;
-use Cake\Event\Event;
 use Cake\I18n\I18n;
 use Cake\View\Helper;
 
@@ -103,22 +102,20 @@ class LibraryHelper extends Helper
 
     /**
      * Before layout callback. beforeLayout is called before the layout is
-     *  rendered.
-     * @param \Cake\Event\Event $event An Event instance
-     * @param string $layoutFile The layout about to be rendered
+     *  rendered
      * @return void
      * @uses MeTools\View\Helper\HtmlHelper::scriptBlock()
      * @uses output
      */
-    public function beforeLayout(Event $event, $layoutFile)
+    public function beforeLayout()
     {
         if (!$this->output) {
             return;
         }
 
         //Writes the output
-        $output = array_map(function ($v) {
-            return "    " . $v;
+        $output = array_map(function ($output) {
+            return "    " . $output;
         }, $this->output);
 
         $this->Html->scriptBlock(
@@ -171,18 +168,17 @@ class LibraryHelper extends Helper
             $scripts[] = '/ckeditor/adapters/jquery';
         }
 
-        //Checks for `APP/webroot/js/ckeditor_init.php`
+        //Checks the init file `APP/webroot/js/ckeditor_init.php` or
+        //  `APP/webroot/js/ckeditor_init.js`.
+        //Otherwise uses the init file `APP/plugin/MeTools/webroot/js/ckeditor_init.js`
+        $init = 'MeTools.ckeditor_init.php?type=js';
         if (is_readable(WWW_ROOT . 'js' . DS . 'ckeditor_init.php')) {
-            $scripts[] = 'ckeditor_init.php?type=js';
-        //Checks for `APP/webroot/js/ckeditor_init.js`
+            $init = 'ckeditor_init.php?type=js';
         } elseif (is_readable(WWW_ROOT . 'js' . DS . 'ckeditor_init.js')) {
-            $scripts[] = 'ckeditor_init';
-        //Else, uses `APP/plugin/MeTools/webroot/js/ckeditor_init.js`
-        } else {
-            $scripts[] = 'MeTools.ckeditor_init.php?type=js';
+            $init = 'ckeditor_init';
         }
 
-        $this->Html->script($scripts, ['block' => 'script_bottom']);
+        $this->Html->script(array_merge($scripts, [$init]), ['block' => 'script_bottom']);
     }
 
     /**
@@ -248,15 +244,14 @@ class LibraryHelper extends Helper
             '/vendor/fancybox/helpers/jquery.fancybox-thumbs',
         ];
 
-        //Checks for `APP/webroot/js/`
-        if (is_readable(WWW_ROOT . 'js' . DS . 'fancybox_init.js')) {
-            $scripts[] = 'fancybox_init';
-        //Else, uses `APP/plugin/MeTools/webroot/fancybox/`
-        } else {
-            $scripts[] = 'MeTools./fancybox/fancybox_init';
+        //Checks the init file inside `APP/webroot/js/`.
+        //Otherwise uses the init file inside `APP/plugin/MeTools/webroot/fancybox/`
+        $init = 'fancybox_init';
+        if (!is_readable(WWW_ROOT . 'js' . DS . 'fancybox_init.js')) {
+            $init = 'MeTools./fancybox/fancybox_init';
         }
 
-        $this->Asset->script($scripts, ['block' => 'script_bottom']);
+        $this->Asset->script(array_merge($scripts, [$init]), ['block' => 'script_bottom']);
     }
 
     /**
@@ -271,7 +266,7 @@ class LibraryHelper extends Helper
      */
     public function shareaholic($siteId)
     {
-        return $this->Html->js('//dsms0mj1bbhn4.cloudfront.net/assets/pub/shareaholic.js', [
+        return $this->Html->script('//dsms0mj1bbhn4.cloudfront.net/assets/pub/shareaholic.js', [
             'async' => 'async',
             'block' => 'script_bottom',
             'data-cfasync' => 'false',