+++ tests/TestCase/View/Helper/HtmlHelperTest.php
@@ -15,7 +15,6 @@ namespace MeTools\Test\TestCase\View\Helper;
 use Cake\Core\Exception\Exception;
 use MeTools\TestSuite\HelperTestCase;
 use MeTools\View\Helper\HtmlHelper;
-use MeTools\View\OptionsParser;
 
 /**
  * HtmlHelperTest class
@@ -49,55 +48,6 @@ class HtmlHelperTest extends HelperTestCase
         $this->expectException(Exception::class);
         $this->expectExceptionMessage('Method `' . HtmlHelper::class . '::noExistingMethod()` does not exist');
         $this->Helper->noExistingMethod(null, null, null);
-    }
-
-    /**
-     * Tests for `addIconToText()` method
-     * @test
-     */
-    public function testAddIconToText()
-    {
-        $text = 'My text';
-
-        $options = optionsParser(['icon' => 'home']);
-        list($result, $options) = $this->Helper->addIconToText($text, $options);
-        $this->assertEquals('<i class="fas fa-home"> </i> ' . $text, $result);
-        $this->assertInstanceOf(OptionsParser::class, $options);
-        $this->assertFalse($options->exists('icon'));
-        $this->assertFalse($options->exists('icon-align'));
-
-        //Missing `icon` option
-        $options = optionsParser(['class' => 'my-class', 'icon-align' => 'right']);
-        list($result, $options) = $this->Helper->addIconToText($text, $options);
-        $this->assertEquals($text, $result);
-        $this->assertInstanceOf(OptionsParser::class, $options);
-        $this->assertFalse($options->exists('icon'));
-        $this->assertFalse($options->exists('icon-align'));
-        $this->assertEquals('my-class', $options->get('class'));
-
-        //Empty text
-        $options = optionsParser(['icon' => 'home']);
-        list($result, $options) = $this->Helper->addIconToText(null, $options);
-        $this->assertEquals('<i class="fas fa-home"> </i>', $result);
-        $this->assertInstanceOf(OptionsParser::class, $options);
-        $this->assertFalse($options->exists('icon'));
-        $this->assertFalse($options->exists('icon-align'));
-
-        //Using `icon-align` option
-        $options = optionsParser(['icon' => 'home', 'icon-align' => 'right']);
-        list($result, $options) = $this->Helper->addIconToText($text, $options);
-        $this->assertEquals($text . ' <i class="fas fa-home"> </i>', $result);
-        $this->assertInstanceOf(OptionsParser::class, $options);
-        $this->assertFalse($options->exists('icon'));
-        $this->assertFalse($options->exists('icon-align'));
-
-        //Invalid `icon-align` option
-        $options = optionsParser(['icon' => 'home', 'icon-align' => 'left']);
-        list($result, $options) = $this->Helper->addIconToText($text, $options);
-        $this->assertEquals('<i class="fas fa-home"> </i> ' . $text, $result);
-        $this->assertInstanceOf(OptionsParser::class, $options);
-        $this->assertFalse($options->exists('icon'));
-        $this->assertFalse($options->exists('icon-align'));
     }
 
     /**
@@ -313,38 +263,6 @@ class HtmlHelperTest extends HelperTestCase
     {
         $expected = $this->Helper->tag('hr', null, ['class' => 'my-hr-class']);
         $this->assertEquals($expected, $this->Helper->hr(['class' => 'my-hr-class']));
-    }
-
-    /**
-     * Test for `icon()` method
-     * @test
-     */
-    public function testIcons()
-    {
-        $expected = ['i' => ['class' => 'preg:/(fa|fab|fal|far|fas) fa\-home/'], ' ', '/i'];
-        foreach ([
-            'home',
-            'fa-home',
-            'fa home',
-            'fas home',
-            'fab home',
-            'fal home',
-            'far home',
-            'fas home',
-            'fa fa-home',
-            ['home'],
-            ['fa', 'home'],
-            ['fas', 'home']
-        ] as $icons) {
-            $this->assertHtml($expected, $this->Helper->icon($icons));
-        }
-        $this->assertHtml($expected, $this->Helper->icon('fa', 'fa-home'));
-
-        $expected = ['i' => ['class' => 'fas fa-hand-o-right fa-2x'], ' ', '/i'];
-        foreach (['hand-o-right 2x', ['hand-o-right', '2x']] as $icons) {
-            $this->assertHtml($expected, $this->Helper->icon($icons));
-        }
-        $this->assertHtml($expected, $this->Helper->icon('hand-o-right', '2x'));
     }
 
     /**
@@ -648,7 +566,7 @@ class HtmlHelperTest extends HelperTestCase
     }
 
     /**
-     * Test for `script()` and `js()` methods
+     * Test for `script()` method
      * @test
      */
     public function testScript()
@@ -656,14 +574,9 @@ class HtmlHelperTest extends HelperTestCase
         //By default, `block` is `true`
         $this->assertNull($this->Helper->script('my-file'));
         $this->assertNull($this->Helper->script('my-file2', ['block' => true]));
-        $this->assertNull($this->Helper->js('my-file4'));
-        $this->assertNull($this->Helper->js('my-file5', ['block' => true]));
 
         $expected = ['script' => ['src' => '/js/my-file3.js']];
         $this->assertHtml($expected, $this->Helper->script('my-file3', ['block' => false]));
-
-        $expected = ['script' => ['src' => '/js/my-file6.js']];
-        $this->assertHtml($expected, $this->Helper->js('my-file6', ['block' => false]));
     }
 
     /**