+++ webroot/js/default.js
@@ -14,12 +14,12 @@
  * Function to auto-submit a form.
  *
  * Example:
- * <select onchange="send_form(this)"></select>
+ * <select onchange="sendForm(this)"></select>
  * @param {Object} element
  */
-function send_form(element)
+function sendForm(element)
 {
-    $(element).closest('form').submit();
+    $(element).closest("form").submit();
 }
 
 /**
@@ -30,12 +30,13 @@ function send_form(element)
  * @see http://stackoverflow.com/a/16788240/1480263
  */
 jQuery.fn.extend({
-    disable: function (state) {
+    disable(state) {
+
         return this.each(function () {
-            if ($(this).is('input, button')) {
-                $(this).prop('disabled', state);
+            if ($(this).is("input, button")) {
+                $(this).prop("disabled", state);
             } else {
-                $(this).toggleClass('disabled', state);
+                $(this).toggleClass("disabled", state);
             }
         });
     }
@@ -44,10 +45,10 @@ jQuery.fn.extend({
 /**
  * Closes a flash messages with an animation
  */
-function close_flashMessage()
+function closeFlashMessage()
 {
-    $('.alert').animate({opacity: '0'}, '800', function () {
-        $(this).slideUp('200');
+    $(".alert").animate({opacity: "0"}, "800", function () {
+        $(this).slideUp("200");
     });
 }
 
@@ -55,21 +56,21 @@ $(function () {
     /**
      * Enables tooltips, if there are elements that require them
      */
-    if ($('[data-toggle~="tooltip"]').length) {
-        $('[data-toggle~="tooltip"]').tooltip();
+    if ($("[data-toggle~=\"tooltip\"]").length) {
+        $("[data-toggle~=\"tooltip\"]").tooltip();
     }
 
     /**
      * Submits button will be disabled when the form is submitted.
      */
-    $('form').submit(function () {
-        $(':submit', this).disable(true);
+    $("form").submit(function () {
+        $(":submit", this).disable(true);
     });
 
     /**
      * Change the visibility of some elements.
      * Elements with "to-be-shown" class will be shown, while elements with "to-be-hidden" class will be hidden
      */
-    $('.hidden.to-be-shown:hidden').removeClass('d-none to-be-shown');
-    $('.to-be-hidden:visible').addClass('d-none').removeClass('to-be-hidden');
+    $(".hidden.to-be-shown:hidden").removeClass("d-none to-be-shown");
+    $(".to-be-hidden:visible").addClass("d-none").removeClass("to-be-hidden");
 });